backend/requirements.txt
cyvcf2
psycopg2
Flask
gunicorn
flask-cors
python-dotenv
